Use Event attributes from schema.org
Schema conforms to schema.org/Event, and perhaps use with Google now
cards. Either way, it's a standard. Note: JSON-ld will be the end
format. If done, remember to add "@" in from of :context and :type
keywords.

This is a fun argument
https://groups.google.com/forum/#!topic/clojure/vM7k9kjOW7E

Also, debate whether the Event schema takes away from a line-oriented
git-diff.